# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuditRequestInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 6 */
  phone: string;
  /** @minLength 1 */
  building: string;
  /** @minLength 3 */
  address: string;
  /** @minLength 10 */
  problem: string;
}

export type AuditRequestUpdateStatus = typeof AuditRequestUpdateStatus[keyof typeof AuditRequestUpdateStatus];


export const AuditRequestUpdateStatus = {
  nou: 'nou',
  acceptat: 'acceptat',
  respins: 'respins',
} as const;

export interface AuditRequestUpdate {
  status?: AuditRequestUpdateStatus;
  formsCollected?: number;
  formsTarget?: number;
  fundsCollected?: number;
  fundsTarget?: number;
}

export type AuditRequestStatus = typeof AuditRequestStatus[keyof typeof AuditRequestStatus];


export const AuditRequestStatus = {
  nou: 'nou',
  acceptat: 'acceptat',
  respins: 'respins',
} as const;

export interface AuditRequest {
  id: number;
  name: string;
  phone: string;
  building: string;
  address: string;
  problem: string;
  status: AuditRequestStatus;
  formsCollected: number;
  formsTarget: number;
  fundsCollected: number;
  fundsTarget: number;
  createdAt: string;
}

export interface AuditSummary {
  total: number;
  thisMonth: number;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface ProjectInput {
  title: string;
  description?: string;
  status?: ProjectStatus;
  imageObjectPath?: string | null;
  beforeImageObjectPath?: string | null;
  afterImageObjectPath?: string | null;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  status: ProjectStatus;
  imageObjectPath?: string | null;
  beforeImageObjectPath?: string | null;
  afterImageObjectPath?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface SpecItemInput {
  title: string;
  description?: string;
  imageObjectPath?: string | null;
  orderNum?: number;
}

export interface SpecItem {
  id: number;
  orderNum: number;
  title: string;
  description: string;
  imageObjectPath?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type DeleteSpecItem200 = {
  ok?: boolean;
};