# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: audit
    description: Free building audit requests
  - name: projects
    description: Projects gallery
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /audit-requests:
    post:
      operationId: createAuditRequest
      tags: [audit]
      summary: Submit a free building audit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditRequestInput"
      responses:
        "201":
          description: Request received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRequest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listAuditRequests
      tags: [audit]
      summary: List submitted audit requests
      responses:
        "200":
          description: Audit requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditRequest"
  /audit-requests/summary:
    get:
      operationId: getAuditSummary
      tags: [audit]
      summary: Get the public audit request summary
      responses:
        "200":
          description: Summary of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditSummary"
  /audit-requests/{id}:
    put:
      operationId: updateAuditRequest
      tags: [audit]
      summary: Update an audit request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuditRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects (public)
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /spec-items:
    get:
      operationId: listSpecItems
      tags: [projects]
      summary: List all specification items (public)
      responses:
        "200":
          description: Spec items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpecItem"
    post:
      operationId: createSpecItem
      tags: [projects]
      summary: Create a spec item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecItem"
  /spec-items/{id}:
    put:
      operationId: updateSpecItem
      tags: [projects]
      summary: Update a spec item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSpecItem
      tags: [projects]
      summary: Delete a spec item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    AuditRequestInput:
      type: object
      required: [name, phone, building, address, problem]
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 6
        building:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 3
        problem:
          type: string
          minLength: 10
    AuditRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [nou, acceptat, respins]
        formsCollected:
          type: number
        formsTarget:
          type: number
        fundsCollected:
          type: number
        fundsTarget:
          type: number
    AuditRequest:
      type: object
      required: [id, name, phone, building, address, problem, status, formsCollected, formsTarget, fundsCollected, fundsTarget, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        phone:
          type: string
        building:
          type: string
        address:
          type: string
        problem:
          type: string
        status:
          type: string
          enum: [nou, acceptat, respins]
        formsCollected:
          type: number
        formsTarget:
          type: number
        fundsCollected:
          type: number
        fundsTarget:
          type: number
        createdAt:
          type: string
          format: date-time
    AuditSummary:
      type: object
      required: [total, thisMonth]
      properties:
        total:
          type: number
        thisMonth:
          type: number
    ProjectInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        imageObjectPath:
          type: string
          nullable: true
        beforeImageObjectPath:
          type: string
          nullable: true
        afterImageObjectPath:
          type: string
          nullable: true
    Project:
      type: object
      required: [id, title, description, status, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [in_progress, completed]
        imageObjectPath:
          type: string
          nullable: true
        beforeImageObjectPath:
          type: string
          nullable: true
        afterImageObjectPath:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SpecItemInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        imageObjectPath:
          type: string
          nullable: true
        orderNum:
          type: number
    SpecItem:
      type: object
      required: [id, title, description, orderNum, createdAt, updatedAt]
      properties:
        id:
          type: number
        orderNum:
          type: number
        title:
          type: string
        description:
          type: string
        imageObjectPath:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string