# lib/api-zod/src/generated/types/auditRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditRequestInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 6 */
  phone: string;
  /** @minLength 1 */
  building: string;
  /** @minLength 3 */
  address: string;
  /** @minLength 10 */
  problem: string;
}