# lib/api-zod/src/generated/types/auditRequestUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditRequestUpdateStatus } from './auditRequestUpdateStatus';

export interface AuditRequestUpdate {
  status?: AuditRequestUpdateStatus;
  formsCollected?: number;
  formsTarget?: number;
  fundsCollected?: number;
  fundsTarget?: number;
}