# lib/api-zod/src/generated/types/specItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpecItem {
  id: number;
  orderNum: number;
  title: string;
  description: string;
  imageObjectPath?: string | null;
  createdAt: Date;
  updatedAt: Date;
}